BuildBot: Add linux-rolling-stable testing
The linux-stable git contains a branch 'linux-rolling-stable' which
gets updates for the most recent Linux release. Instead of adding
separate "short-term" stable branches, this should do the same work
unattended.

Normal handling of the branches in linux-stable git uses a kernel
configuration file created for the specific LTS stable versions. The
linux-rolling-stable branch is close to the current Linux mainline so
the kernel configuration file for the Linux mainline should be used as a
base for linux-rolling-stable kernel.

Add exceptions to normal stable handling that a 'linux-rolling-stable'
configuration file should not be used when that is the linux-stable
kernel being tested.

Add exception to use mainline kernel config if the linux-stable kernel
being tested is from the 'linux-rolling-stable' branch.

----
Additional in this update: Hide git bisect related BuildBot steps when
they are not executed.