BuildBot / LabGrid: Change 'pmic' test dir name
The test system has tests for other component types than PMICs. Original
thought was that when more tests are needed, the programming needs to be
more diverge that what it ended up being. This is why the directory
containing driver tests is named 'pmic' and the sensor tests are
included in that directory as well.

Rename 'pmic' directory to 'driver_tests'.

Change python files that work on that directory to use 'driver_tests'
instead of 'pmic'.

----
Additionally delete leftover files that are no longer needed, some of
which ended up to remote repository accidentally.