Tests: ADDAC: Read ADC 10x to get smoother input
Noise causes some small value jumps and we do not have analog filter on
the input. To solve this problem and to be able to get smaller tolerance
value on readings, smooth out the noise by reading 10x values and divide
by 10.

Shell script is used to read the values. This is much faster than
opening UART multiple times to do the readings.

Additionally this commit renames 'test_init' to get_sysfs_information.