BD79124: Change regulator voltage in DT
The device tree voltage that BD71924 is receiving is defined at 3.3V.
However, the BeagleBone Black's 3.3V output is a bit more than that,
measured at 3.4V with multimeter.

This might be the cause that the voltage value calculated back from ADC
input and multiplier starts to fall off at higher readings.