Test System: Add 'test_linux' config.
The test linux branch in Linux-Driver-Testing is used in testing stage
of developing new features to the test system, including new tests.

Some of the drivers for which tests that are being developed are not
supported in any public kernel release yet, but are only in the stage
of being accepted or still developed.

Add 'test_linux' specific configuration file. This config has enabled
BD96801/2/5/6
BD79700/1/2/3
BD79124

Head of the Linux-Driver-Testing/test_linux has a kernel patched with
the driver versions that are now accepted.