BuildBot: Fix linux-rolling-stable test launching
The commit message for the commit that should be built is a merge
commit, with the topic "Merge v##.##.###".

The regular stable scheduler wont work with that, and the builds are not
launched.

Add a function doing separate regex for linux-rolling-stable commit
matching.

Add a get function for adding linux-rolling-stable scheduler with the
above function.

Branch adding schedulers with if statement for LTS stable and rolling
stable.