Tests: Move RTC functionaliy from PMIC class to RTC class
It is better to have the RTC test functionality in its own test class
instead of PMIC class. Other components should be able to use the same
functionality, and if not, it should not be too hard to fix.

Move all RTC related functions and info structures from pmic_class.py to
rtc_class.py