DTS: Modify the _test.dts to not describe spi0 bus
The automated driver test system will be modified to install AD/DA
converters to the system in two phases - 1. overlay a devicetree
describing the SPI node. 2. describe the components that are to be
tested.

Modify devicetrees to only describe the installed component to spi0
node.

The original devicetrees hold some value for manual testing, so they are
included here.