Kernel module: Fix overlay merger compatibility for next
There is currently a change in linux-next which requires changes to
overlay merger module. The change is expected to be included in Linux
6.16, but linux-next kernel version is 6.15-rc#.

This change causes compilation error and needs to be fixed in order to
have the tests functional.

The kernel version macros needs to be changed once this change ends up
in the mainline release.

----
Patch in question:
https://lore.kernel.org/all/20250313-sysfs-const-bin_attr-final-v2-0-96284e1e88ce@weissschuh.net/