Facotiers: Separate kernel config and test modules for linux-next
Out of the tested Linux kernel versions, linux-next is the first place
where support for new ROHM drivers appears. It is also the first place
where changes affecting the test system's out-of-tree kernel modules
appear.

This is excellent spot to start introducing bug fixes to kernel modules
used in the test system, and also to start automated driver tests for
newly introduced drivers. To do so, separate kernel config is required
that enables support for drivers that are not yet in the mainline Linux
kernel release. The branch of test-kernel-modules used for linux-next is
specific to linux-next, and can include updates to test modules as
needed.