Test results: Add information about tests
To make results more powerful, add information to results about what the
failed test is actually doing.

Additional fix here is to check if the test stage is installing the test
modules, as the results are written in the same function as merging
device tree overlay.

Additionally, removed lines of code taht are not in use.