Disable tests for BD79701 for now
There is a random unresolved bug happening sometimes when reading
BD79104 channel. BD79701 DAC is tested with the BD79104 ADC.

When the bug does not happen, tests are passing. But when the bug
happens, it causes immediate kernel crash.

The root cause of the bug is unknown for now. The likely places to look
for the bug is the BeagleBones SPI driver and Linux SPI core. The
BD79104 driver is unlikely place.

Both of these ICs will get included back in the setup once the bug is
resolved. Keeping them in the test system invalidates both, passed and
failed results in the top level -  failures might be kind of OK if the
only thing failed is the BD79701 / BD79104 test.